Update progress bar to limit bar width to between 0 and 100%
UtilizationBarChart has been updated such that the used bar width
will never render over 100%.  At a bar width >100%, the overflow
is clipped by the parent div.  However, the tooltip render does
not factor in the clipped region causing the tooltip text to render
far to the right of the intended area.

Limiting the max width of the used area to 100% avoids having to
make any changes to the patternfly or bootstrap code itself.

Change-Id: Ie1813e46802f74d585f84afc518dc7c990dbfeba
Bug-Url: https://bugzilla.redhat.com/1368768
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>