Convert intl.js Date/Time formatting to ECMA-402 Intl API with UTC TZ
By using the Intl API, we gain discrete control over date/time
formatting.  This also allows the formatted strings to have
locale specific formats and names.  The formatted strings are
also rendered in the UTC time zone.  This will allow for consistency
with the oVirt event log when it is converted to UTC time.

Caveat: Current implementations of the Intl API in the various
browsers don't support more then the ECMA-402 defined 2 default
time zones of "runtime default" and "UTC".  Chrome is currently
the only browser that allow you to use more then the minimum
required 2 time zones.

If, in future, we want the dashboard to render date/times in a TZ
other then browser default or UTC, we will most likely need to use a
non-Intl API formatter.

Change-Id: I115a07af7f826fe85fbeb2ea17621bcc4e6d538c
Bug-Url: https://bugzilla.redhat.com/1362402
Bug-Url: https://bugzilla.redhat.com/1356909
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>