Adjust content wrapping in aggregate status cards
On the aggregate status card, with some combinations of large counts
of all statuses, it could happen that content would wrap to the next
line.  In even more specific circumstances, the status icon could be
on a different line from the count number.

This CSS change uses new flex box layouts to wrap the content
appropriately while keeping the icon and count together.

Change-Id: Icebf0431c1636252baff5a1b95f73c786fd7c7c4
Bug-Url: https://bugzilla.redhat.com/1392984
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>