Fix tooltip positioning on SVG elements when the page is scrolled
HeatMap charts on the dashboard use SVG to render. When the page was
scrolled, tooltips did not render the correct position.  Bootstrap
3.3.7 introduced a change to the tooltip positioning code
specifically for SVG elements due to problems with jQuery.offset()
and SVG elements. Since the problem won't be addressed by maintainers
of jQuery or Bootstrap, the problem has been fixed in a dashboard
copy of bootstrap tooltip code.

Bootstrap version has been locked to 3.3.7 to force any updates to
future 3.x releases to be a manual process.  This will help ensure
the tooltip position override code remains in sync with bootstrap
code.

Note: angular-patternfly does not have the same issue with tooltips
on their HeatMap implementation.  angular-patternfly is based on
angular-ui-bootstrap which doesn't use jQuery for position
calculations.

Change-Id: I0a53b65d05a2be5f0620618818ec762b19c58f22
Bug-Url: https://bugzilla.redhat.com/1365929
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>