intl - Use rounding for numbers passed to message formats
Converted format message values from formatted number strings
to rounded numbers so IntlMessageFormat can localize the
numbers itself.  Giving a localize number to the message
formatter will cause a parse error if the localized format
can't be coerced to a number. For example, '1.234,56' can't be
coerced to the value 1234.56.

Two new number styles were added so that message formats can
specify directly if a number is to be an integer with rounding
(e.g. "{total, number, 0}") or a single decimal place with
rounding (e.g. "{total, number, 0.0}").

Additional unit test for the rounding function and the message
format number styles have been added for validation and examples
of functionality.

Change-Id: I3f7c01ccb7a79740312104b0535b5c5d66b5ae13
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1421104
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1489152
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>