Make all props of GlobalDashboard required
<GlobalDashboard> component needs all of its props to work
and display correctly. When <DashboardDataProvider> renders,
it takes its child and clones it with all props required by
the <GlobalDashboard> component.

Without <DashboardDataProvider> component (e.g. if we decide
to showcase our presentational/non-infra UI components using
tools like Playbook), we'd still need to pass all props to
the <GlobalDashboard> component for it to work properly.

Change-Id: I0f7263772965cf28f3221791de283e3e49808780
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>