Hide tooltip on refresh button when it's clicked or expanded
The bootstrap tooltip default trigger is on hover and focus. When
the tooltip is on top of a button, or a button group, it will
therefore stay visible when the button is clicked or a group is
expanded.  Changing the tooltip to hover only and attaching a click
handler to the button itself to hide the tooltip works well so
a click or group expand hides it.

Change-Id: Ife7174c83156c0934465bdd3631ec28e60603078
Bug-Url: https://bugzilla.redhat.com/1510524
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>