Add text length resize sensing on status cards for height matching
Under certain conditions, the webfont used by PF/RCUE will not be
loaded to the browser prior to the aggregate status cards getting
their content height matched.  Sometimes the heights are matched
and the status strings change length once the icons are loaded and
the text wraps.  The height match sets a static height and the text
wraps below the component.  This only happens consistently when
the first status card body wraps but none of the other cards do.

The fix could can be any of:
  (1) make the font load faster
  (2) wait until the fonts are loaded before rendering the app
  (3) load the app but explicitly fire the height matching again
      once the fonts are loaded
  (4) use element resize/reflow detection that will fire once the
      font is loaded and the icons take up their expected space

However:
  (1) Even if inlined as URL data by webpack, the fonts still may
      not load fast enough on a slow browser.
  (3) If we retrigger on load, there could be a flash of unstyled
      content (FOUC).  That is to be avoided if possible.

Options (2) and (4) are viable options:
  (2) Any font based FOUC in the app would be prevented, but the app
      will be delayed until the fonts are all fully loaded (50-60ms
      extra on a development environment but done in parallel to the
      initial data fetch which would probably take longer in 50-60ms).
  (4) Works as long as the resize detect is attached in the gap between
      the initial height matching and the text length changing because
      the font gets loaded.  This option has the added benefit of
      also handling window resize detection.

This patch implements option (4) as it cleanly covers this issue and
wraps in handling for some other cases as well.

Change-Id: I8bd1885334ec85eb26c3ab80070b67e00bbe5b54
Bug-Url: https://bugzilla.redhat.com/1489467
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>