Use Patternfly from official npm package
Removed the statically embedded copy of rcue and replace it with an
import from the patternfly npm module.  Going forward, a version
number change and rebuild should be enough to upgrade to a new version
of patternfly, assuming the CSS changes are non-breaking.

Patternfly and bootstrap are referenced at a static version to prevent
automatic updates of those libraries because of semver version ranges.

Webpack build was updated to allow importing of CSS in the jsx
files.  The base HTML files are now generated and get both the
generated js and css files injected on the page (at build time).

Change-Id: I17120742b897d4a068d7e31d7f31d0725b41d9a5
Bug-Url: https://bugzilla.redhat.com/1451485
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>